Overhaul subset sum calculations
The prevous implementation wasn't taking into consideration the fact
that different cells could have a reduced set of possible values.

That is, given the following cage with sum 10:
  +---+---+
  |   |123|
  | 1 |456|
  |   |789|
  +---+---+

Running killer combos or subset sums on this cage would return the
following: { [1,9], [2,8] [3,7], [4,6] }

... where each value in each subset could be found in each cell. The result
would therefore be:
  +---+---+
  |   |123|
  | 1 |4 6|
  |   |789|
  +---+---+

Even though the more accurate result would be a solved cage:
  +---+---+
  |   |   |
  | 1 |   |
  |   |  9|
  +---+---+

This fixes that up to provide more accurate results.