Add a step to expose hidden pairs in two-cell cages
Might be a bit specific, but is fairly interesting:

    0   1   2
  +---+---+---+
  |   |   |   |
H |   |   |45 |
  |   |   |78 |
  +---+---| - |
  |   |   |   |
J |   |   |45 |
  |   |   |78 |
  +---+---+---+

Say 8 is only defined by H2/J2 in a column/box. Since they share a cage,
we know that 8 must be found in one of these. The other cage would then
be a 4 (assuming the sum is 12). We can then set these two cells to 4/8,
and take advantage of the naked pair.