Optimize the subset-sum generation algorithm
This moves things around a bit and adds some more logic to avoid
unnecessary recursion. Most of these optimizations are reliant on the
fact that the list of candidates/possibles are ordered.

It also uses a couple of extra parameters to avoid the calculation of
the subset sum and possibles' list's size.

This improves performance considerably for the typical case. I've seen
improvements of over 50%.

For cases where the added checks are rarely taken then it doesn't make
things any better, sadly.