Refactor cage combos
Initialize them with the grid and store them. Then, when a step modifies
anything, go through the cage combo map and remove the ones that are
invalid.

This saves us the re-computation of cage combos every time the killer
combos or cage/unit overlap steps are run, which leads to some nice
performance improvements.