Add an initial version of loading sudokus from files
The format of the file is yet to be documented, but it's essentially the
following:

First, 9 lines of 9 hexadecimal numbers. These specify the candidate
values for each of the 81 cells. The hexadecimal numbers are hex
versions of binary numbers, where each digit represents a cell candidate
value.

1FF being {1/2/3/4/5/6/7/8/9}, 3 being {1/2}, 1 being {1}, etc.

Then comes the cage values. The format is the following:

    cage_sum <cells>

The 'cage_sum' field is an integer.

The <cells> field lists the cells in the cage, each separated by a
space. There are some space-saving features, too. Examples:

    9 A0 B0 C0  # Sum: 9, Cells: A0, B0, C0
    9 ABC0      # Same

    8 AB0 BC1   # Sum: 8, Cells: A0, B0, B1, C1

    6 A123      # Sum: 6, Cells: A1, A2, A3