Add basic support for custom strategies
The idea is that the user can specify the specific strategy they want to
run on the grid. A strategy consists of blocks of steps to run. A block
can also be repeated a given number of times, forming a loop.

A loop iteration detects whether it has made any progress, and will bail
if it hasn't, preventing it from running longer than it needs to.

The current functionality only mirrors that of the previous pipeline in
that the user either runs a single step, or the default set. In the
future, the user will be able to provide their own via a file and some
custom syntax.