Add a minimal testing framework inspired by LLVM LIT
This adds `run_tests.py`, which can run tests as defined in `#` comments
at the beginning of test files. They can run arbitrary `columbo`
commands, and may optionally check their output against
`columbo_check.py`, which is a diff-like utility. It just strips
comments for now, but can be used to regression test individual steps.

A summary of passed, failed, unsupported, and skipped tests is printed
at the end of the run.

A mix of test files and paths can be passed to `run_tests.py`, which are
collected as the suite to run. Else a set of default paths will be run.

Tests are defined as `# RUN:` commands in the first line(s) of a file.
They must be a command separated by pipes (`|`) and each component must
something called either `columbo` or `columbo_check` which are replaced
by the full paths to the respective commands. Substitutions of `%s` and
`%S` are supported to substitute paths to the file path or file
directory path, respectively.