Use innie/outie cages in conflicting-combo strategies
This makes innie/outie cages more permanent by registering them and
storing them in the InnieOutieRegion. They are treated as pseudo-cages
by their component cells. Often, innies and outies that align with
rows/columns/boxes can be used to reduce the combination sets of cages
that share those nonnets.

Case in point: this new krazydad which we can now solve!